Sprint 16.1 разнесены Settings/Sharing (CleanArch)
