+++ app/build.gradle
@@ -54,4 +54,5 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
+    implementation 'androidx.core:core-ktx:1.9.0'
 }
