+++ app/src/main/AndroidManifest.xml
@@ -18,23 +18,23 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.LibraryActivity"
+            android:name=".LibraryActivity"
             android:exported="false"
             />
         <activity
-            android:name=".presentation.ui.SettingsActivity"
+            android:name=".settings.ui.SettingsActivity"
             android:exported="false"
            />
         <activity
-            android:name=".presentation.ui.MediaActivity"
+            android:name=".player.ui.MediaActivity"
             android:exported="false"
            />
         <activity
-            android:name=".presentation.ui.SearchActivity"
+            android:name=".search.SearchActivity"
             android:exported="false"
             />
         <activity
-            android:name=".presentation.ui.MainActivity"
+            android:name=".main.ui.MainActivity"
             android:exported="true"
             >
