+++ app/src/main/java/com/example/playlistmaker/LibraryActivity.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.playlistmaker.R
+
+class LibraryActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_library)
+    }
+}
