--- app/src/main/java/com/example/playlistmaker/domain/Player.kt
@@ -1,17 +1,0 @@
-package com.example.playlistmaker.domain
-
-import com.example.playlistmaker.data.PlayerImpl
-
-interface Player {
-    fun startAudio()
-    fun pauseAudio()
-    fun isPlaying(): Boolean
-    fun currentPosition(): Int
-    fun playbackControl(onStartPlayer: () -> Unit, onPausePlayer: () -> Unit)
-    fun preparePlayer(
-        dataSource: String,
-        onPreparedListener: () -> Unit,
-        onCompletionListener: () -> Unit
-    )
-}
-
