--- app/src/main/java/com/example/playlistmaker/domain/PlayerInteractor.kt
@@ -1,15 +1,0 @@
-package com.example.playlistmaker.domain
-
-interface PlayerInteractor {
-    fun startAudio()
-    fun pauseAudio()
-    fun isPlaying(): Boolean
-    fun currentPosition(): Int
-    fun playbackControl(onStartPlayer: () -> Unit, onPausePlayer: () -> Unit)
-
-    fun preparePlayer(
-        dataSource: String,
-        onPreparedListener: () -> Unit,
-        onCompletionListener: () -> Unit
-    )
-}
