+++ app/src/main/java/com/example/playlistmaker/player/domain/api/MediaRepository.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.player.domain.api
+
+interface MediaRepository {
+    fun getTrackCoverUrl(): String?
+    fun getTrackName(): String?
+    fun getArtistName(): String?
+    fun getTrackTime(): String?
+    fun getCollectionName(): String?
+    fun getReleaseDate(): String?
+    fun getPrimaryGenreName(): String?
+    fun getCountry(): String?
+    fun getPreviewUrl(): String?
+}
