+++ app/src/main/java/com/example/playlistmaker/player/domain/api/PlayerInteractor.kt
@@ -1,0 +1,15 @@
+package com.example.playlistmaker.player.domain.api
+
+interface PlayerInteractor {
+    fun startAudio()
+    fun pauseAudio()
+    fun isPlaying(): Boolean
+    fun currentPosition(): Int
+    fun playbackControl(onStartPlayer: () -> Unit, onPausePlayer: () -> Unit)
+
+    fun preparePlayer(
+        dataSource: String,
+        onPreparedListener: () -> Unit,
+        onCompletionListener: () -> Unit
+    )
+}
