+++ app/src/main/java/com/example/playlistmaker/player/ui/MediaContract.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.player.ui
+
+interface MediaContract {
+    interface Presenter {
+        fun onPlayClicked()
+        fun onPauseAudioClicked()
+        fun onFavoriteClicked()
+        fun onDisLikeClicked()
+        fun onPause()
+
+    }
+}
