--- app/src/main/java/com/example/playlistmaker/presentation/ui/LibraryActivity.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker.presentation.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.playlistmaker.R
-
-class LibraryActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_library)
-    }
-}
