--- app/src/main/java/com/example/playlistmaker/presentation/ui/MainActivity.kt
@@ -1,49 +1,0 @@
-package com.example.playlistmaker.presentation.ui
-
-import android.annotation.SuppressLint
-import android.content.Context
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Button
-import androidx.appcompat.app.AppCompatDelegate
-import com.example.playlistmaker.R
-
-class MainActivity : AppCompatActivity() {
-
-    @SuppressLint("MissingInflatedId")
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setAppTheme()
-        setContentView(R.layout.activity_main)
-
-        val btnSearch = findViewById<Button>(R.id.btnSearch)
-        btnSearch.setOnClickListener {
-            val intent = Intent(this, SearchActivity::class.java)
-            startActivity(intent)
-        }
-
-        val btnLibrary = findViewById<Button>(R.id.btnLibrary)
-        btnLibrary.setOnClickListener {
-            val intent = Intent(this, LibraryActivity::class.java)
-            startActivity(intent)
-        }
-
-        val btnSettings = findViewById<Button>(R.id.btnSettings)
-        btnSettings.setOnClickListener {
-            val intent = Intent(this, SettingsActivity::class.java)
-            startActivity(intent)
-        }
-
-    }
-
-    private fun setAppTheme() {
-        val sharedPrefs = getSharedPreferences("settings", Context.MODE_PRIVATE)
-        val darkMode = sharedPrefs.getBoolean("darkTheme", false)
-        if (darkMode) {
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-        } else {
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-        }
-    }
-}
