--- app/src/main/java/com/example/playlistmaker/presentation/ui/MediaContract.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker.presentation.ui
-
-interface MediaContract {
-    interface Presenter {
-        fun onPlayClicked()
-        fun onPauseAudioClicked()
-        fun onFavoriteClicked()
-        fun onDisLikeClicked()
-        fun onPause()
-
-    }
-}
