+++ app/src/main/java/com/example/playlistmaker/settings/data/SettingsRepositoryImpl.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.settings.data
+
+import android.content.SharedPreferences
+import com.example.playlistmaker.settings.domain.SettingsRepositoryInterface
+
+class SettingsRepositoryImpl(private val sharedPrefs: SharedPreferences) : SettingsRepositoryInterface {
+    override fun getDarkTheme(): Boolean {
+        return sharedPrefs.getBoolean("darkTheme", false)
+    }
+
+    override fun setDarkTheme(enabled: Boolean) {
+        sharedPrefs.edit().putBoolean("darkTheme", enabled).apply()
+    }
+}
