+++ app/src/main/java/com/example/playlistmaker/settings/domain/SettingsRepositoryInterface.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.settings.domain
+
+interface SettingsRepositoryInterface {
+    fun getDarkTheme(): Boolean
+    fun setDarkTheme(enabled: Boolean)
+}
