+++ app/src/main/java/com/example/playlistmaker/settings/domain/SettingsUseCaseInterface.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.settings.domain
+
+interface SettingsUseCaseInterface {
+    fun getDarkTheme(): Boolean
+    fun setDarkTheme(enabled: Boolean)
+}
