+++ app/src/main/java/com/example/playlistmaker/sharing/domain/SharingRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.sharing.domain
+
+interface SharingRepository {
+    fun shareApp(text: String)
+    fun sendSupportEmail(subject: String, body: String)
+    fun openAgreementUrl(text: String)
+}
