+++ app/src/main/java/com/example/playlistmaker/sharing/domain/SharingUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.sharing.domain
+
+interface SharingUseCase {
+    fun shareApp()
+    fun sendSupportEmail()
+    fun openAgreementUrl()
+}
