+++ app/src/main/res/layout/activity_library.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.LibraryActivity">
+    tools:context=".LibraryActivity">
 
     <ImageView
         android:layout_width="50dp"
